improve support for MCC modern syntax
Now parsing ReachabilityFireability queries correctly.

we use AG (in two nodes) and EF at top of MCC current format for
properties.

There are no "tags" present either, so we disable that test when parsing
E or A as first formula node.

Also patch strange "-" characters in the file that were messing up the
parser.